requirements.txt
acres==0.5.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
cleo==2.1.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
etelemetry==0.3.1
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
fitz==0.0.1.dev2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
groq==0.29.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keyring==25.6.0
langchain-core==0.3.66
langchain-groq==0.3.4
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
more-itertools==10.7.0
msgpack==1.1.1
narwhals==1.44.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
oauthlib==3.3.1
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
pbs-installer==2025.6.26
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry==2.1.3
poetry-core==2.1.3
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.29
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.26.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
scipy==1.16.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.46.1
streamlit-aggrid==1.1.6
tenacity==9.1.2
tinydb==4.8.2
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.1
traits==7.0.2
trove-classifiers==2025.5.9.12
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
watchdog==6.0.0
zstandard==0.23.0